# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CompanySettings {
  name: string;
  cnpj: string;
  phone: string;
  email: string;
  address: string;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  logoText?: string | null;
}

export interface CompanySettingsInput {
  name: string;
  cnpj: string;
  phone: string;
  email: string;
  address: string;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  logoText?: string | null;
}

export interface Client {
  id: string;
  name: string;
  document: string;
  phone: string;
  email: string;
  address: string;
  createdAt: string;
}

export interface ClientInput {
  name: string;
  document: string;
  phone: string;
  email: string;
  address: string;
}

export interface Product {
  id: string;
  name: string;
  description: string;
  price: number;
  unit: string;
}

export interface ProductInput {
  name: string;
  description: string;
  price: number;
  unit: string;
}

export interface Service {
  id: string;
  name: string;
  description: string;
  price: number;
  /** @nullable */
  category?: string | null;
}

export interface ServiceInput {
  name: string;
  description: string;
  price: number;
  /** @nullable */
  category?: string | null;
}

export interface DocumentClientInfo {
  name: string;
  document: string;
  phone: string;
  email: string;
  address: string;
}

export type DocumentItemType = typeof DocumentItemType[keyof typeof DocumentItemType];


export const DocumentItemType = {
  produto: 'produto',
  servico: 'servico',
} as const;

export interface DocumentItem {
  id: string;
  name: string;
  type: DocumentItemType;
  quantity: number;
  price: number;
  total: number;
}

export type DocumentType = typeof DocumentType[keyof typeof DocumentType];


export const DocumentType = {
  orcamento: 'orcamento',
  ordem_servico: 'ordem_servico',
  laudo_tecnico: 'laudo_tecnico',
  recibo: 'recibo',
} as const;

export interface Document {
  id: string;
  type: DocumentType;
  number: string;
  date: string;
  /** @nullable */
  dueDate?: string | null;
  client: DocumentClientInfo;
  items: DocumentItem[];
  totalAmount: number;
  /** @nullable */
  discount?: number | null;
  /** @nullable */
  observations?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  equipment?: string | null;
  /** @nullable */
  serialNumber?: string | null;
  /** @nullable */
  defect?: string | null;
  /** @nullable */
  solution?: string | null;
  /** @nullable */
  technicalAnalysis?: string | null;
  /** @nullable */
  conclusion?: string | null;
  /** @nullable */
  responsavelTecnico?: string | null;
  /** @nullable */
  registroProfissional?: string | null;
  /** @nullable */
  receivedValue?: number | null;
  /** @nullable */
  referringTo?: string | null;
  /** @nullable */
  issuerName?: string | null;
  /** @nullable */
  issuerDocument?: string | null;
  createdAt?: string;
}

export type DocumentInputType = typeof DocumentInputType[keyof typeof DocumentInputType];


export const DocumentInputType = {
  orcamento: 'orcamento',
  ordem_servico: 'ordem_servico',
  laudo_tecnico: 'laudo_tecnico',
  recibo: 'recibo',
} as const;

export interface DocumentInput {
  /** @nullable */
  id?: string | null;
  type: DocumentInputType;
  number: string;
  date: string;
  /** @nullable */
  dueDate?: string | null;
  client: DocumentClientInfo;
  items: DocumentItem[];
  totalAmount: number;
  /** @nullable */
  discount?: number | null;
  /** @nullable */
  observations?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  equipment?: string | null;
  /** @nullable */
  serialNumber?: string | null;
  /** @nullable */
  defect?: string | null;
  /** @nullable */
  solution?: string | null;
  /** @nullable */
  technicalAnalysis?: string | null;
  /** @nullable */
  conclusion?: string | null;
  /** @nullable */
  responsavelTecnico?: string | null;
  /** @nullable */
  registroProfissional?: string | null;
  /** @nullable */
  receivedValue?: number | null;
  /** @nullable */
  referringTo?: string | null;
  /** @nullable */
  issuerName?: string | null;
  /** @nullable */
  issuerDocument?: string | null;
}