# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: company
    description: Company settings
  - name: clients
    description: Client management
  - name: products
    description: Product catalog
  - name: services
    description: Service catalog
  - name: documents
    description: Document management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /company-settings:
    get:
      operationId: getCompanySettings
      tags: [company]
      summary: Get company settings
      responses:
        "200":
          description: Company settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanySettings"
    put:
      operationId: updateCompanySettings
      tags: [company]
      summary: Update company settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanySettingsInput"
      responses:
        "200":
          description: Updated company settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanySettings"

  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"

  /clients/{id}:
    put:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/{id}:
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all services
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      summary: Create a service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Created service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"

  /services/{id}:
    put:
      operationId: updateService
      tags: [services]
      summary: Update a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "200":
          description: Updated service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "404":
          description: Not found
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List all documents
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Create a document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Created document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"

  /documents/{id}:
    put:
      operationId: updateDocument
      tags: [documents]
      summary: Update a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "200":
          description: Updated document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    CompanySettings:
      type: object
      required: [name, cnpj, phone, email, address]
      properties:
        name:
          type: string
        cnpj:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        website:
          type: ["string", "null"]
        logoText:
          type: ["string", "null"]

    CompanySettingsInput:
      type: object
      required: [name, cnpj, phone, email, address]
      properties:
        name:
          type: string
        cnpj:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        website:
          type: ["string", "null"]
        logoText:
          type: ["string", "null"]

    Client:
      type: object
      required: [id, name, document, phone, email, address, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        document:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        createdAt:
          type: string

    ClientInput:
      type: object
      required: [name, document, phone, email, address]
      properties:
        name:
          type: string
        document:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string

    Product:
      type: object
      required: [id, name, description, price, unit]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        unit:
          type: string

    ProductInput:
      type: object
      required: [name, description, price, unit]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        unit:
          type: string

    Service:
      type: object
      required: [id, name, description, price]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: ["string", "null"]

    ServiceInput:
      type: object
      required: [name, description, price]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: ["string", "null"]

    DocumentClientInfo:
      type: object
      required: [name, document, phone, email, address]
      properties:
        name:
          type: string
        document:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string

    DocumentItem:
      type: object
      required: [id, name, type, quantity, price, total]
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [produto, servico]
        quantity:
          type: number
        price:
          type: number
        total:
          type: number

    Document:
      type: object
      required: [id, type, number, date, client, items, totalAmount]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [orcamento, ordem_servico, laudo_tecnico, recibo]
        number:
          type: string
        date:
          type: string
        dueDate:
          type: ["string", "null"]
        client:
          $ref: "#/components/schemas/DocumentClientInfo"
        items:
          type: array
          items:
            $ref: "#/components/schemas/DocumentItem"
        totalAmount:
          type: number
        discount:
          type: ["number", "null"]
        observations:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        equipment:
          type: ["string", "null"]
        serialNumber:
          type: ["string", "null"]
        defect:
          type: ["string", "null"]
        solution:
          type: ["string", "null"]
        technicalAnalysis:
          type: ["string", "null"]
        conclusion:
          type: ["string", "null"]
        responsavelTecnico:
          type: ["string", "null"]
        registroProfissional:
          type: ["string", "null"]
        receivedValue:
          type: ["number", "null"]
        referringTo:
          type: ["string", "null"]
        issuerName:
          type: ["string", "null"]
        issuerDocument:
          type: ["string", "null"]
        createdAt:
          type: string

    DocumentInput:
      type: object
      required: [type, number, date, client, items, totalAmount]
      properties:
        id:
          type: ["string", "null"]
        type:
          type: string
          enum: [orcamento, ordem_servico, laudo_tecnico, recibo]
        number:
          type: string
        date:
          type: string
        dueDate:
          type: ["string", "null"]
        client:
          $ref: "#/components/schemas/DocumentClientInfo"
        items:
          type: array
          items:
            $ref: "#/components/schemas/DocumentItem"
        totalAmount:
          type: number
        discount:
          type: ["number", "null"]
        observations:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        equipment:
          type: ["string", "null"]
        serialNumber:
          type: ["string", "null"]
        defect:
          type: ["string", "null"]
        solution:
          type: ["string", "null"]
        technicalAnalysis:
          type: ["string", "null"]
        conclusion:
          type: ["string", "null"]
        responsavelTecnico:
          type: ["string", "null"]
        registroProfissional:
          type: ["string", "null"]
        receivedValue:
          type: ["number", "null"]
        referringTo:
          type: ["string", "null"]
        issuerName:
          type: ["string", "null"]
        issuerDocument:
          type: ["string", "null"]