# lib/api-zod/src/generated/types/companySettingsInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CompanySettingsInput {
  name: string;
  cnpj: string;
  phone: string;
  email: string;
  address: string;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  logoText?: string | null;
}