# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentClientInfo } from './documentClientInfo';
import type { DocumentItem } from './documentItem';
import type { DocumentType } from './documentType';

export interface Document {
  id: string;
  type: DocumentType;
  number: string;
  date: string;
  /** @nullable */
  dueDate?: string | null;
  client: DocumentClientInfo;
  items: DocumentItem[];
  totalAmount: number;
  /** @nullable */
  discount?: number | null;
  /** @nullable */
  observations?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  equipment?: string | null;
  /** @nullable */
  serialNumber?: string | null;
  /** @nullable */
  defect?: string | null;
  /** @nullable */
  solution?: string | null;
  /** @nullable */
  technicalAnalysis?: string | null;
  /** @nullable */
  conclusion?: string | null;
  /** @nullable */
  responsavelTecnico?: string | null;
  /** @nullable */
  registroProfissional?: string | null;
  /** @nullable */
  receivedValue?: number | null;
  /** @nullable */
  referringTo?: string | null;
  /** @nullable */
  issuerName?: string | null;
  /** @nullable */
  issuerDocument?: string | null;
  createdAt?: string;
}