# lib/api-zod/src/generated/types/documentClientInfo.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DocumentClientInfo {
  name: string;
  document: string;
  phone: string;
  email: string;
  address: string;
}